Выполняет разбиение кода на модули
