+++ js/data.js
@@ -1,0 +1,76 @@
+import {getRandomInteger, getRandomArrayElement, createIdGenerator} from './util.js';
+
+const DESCRIPTIONS = [
+  'Лучшее фото в моей жизни!',
+  'Это прекрасно!',
+  'Лепота-то какая!',
+  'Мне нравится',
+  'Думаю, красиво',
+  'Очень хорошо сегодня',
+  'Чудесный был день!',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Игорь',
+  'Ян',
+  'Роберт',
+  'Юрий',
+  'Аркадий',
+  'Лена',
+  'Татьяна',
+  'Илья',
+  'Фёдор',
+];
+
+const SIMILAR_PHOTO_COUNT = 25;
+const SIMILAR_COMMENT_COUNT = 3;
+
+
+const createRandomId = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+const generateCommentId = createIdGenerator();
+const generatePhotoId = createIdGenerator();
+const generateImageId = createIdGenerator();
+const generateLikesQuantity = createRandomId(15, 200);
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(MESSAGES),
+  name: getRandomArrayElement(NAMES),
+});
+
+const createPhoto = () => ({
+  id: generatePhotoId(),
+  url: `photos/${generateImageId()}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: generateLikesQuantity(),
+  comments: Array.from({ length: SIMILAR_COMMENT_COUNT }, createComment),
+});
+
+const createPhotos = () => Array.from({ length: SIMILAR_PHOTO_COUNT }, createPhoto);
+
+export {createPhotos};
